+++ l10n/pl.js
@@ -64,6 +64,7 @@ OC.L10N.register(
     "You are using public demo ONLYOFFICE Document Server. Please do not store private sensitive data." : "Używasz publicznego serwera demo. Nie przechowuj tu prywatnych danych.",
     "Select tag" : "Wybierz tag",
     "version": "wersja",
-    "Disable certificate verification (insecure)": "Wyłącz weryfikację certyfikatu (niebezpieczne)"
+    "Disable certificate verification (insecure)": "Wyłącz weryfikację certyfikatu (niebezpieczne)",
+    "Generate document preview": "Wygeneruj podgląd dokumentu"
 },
 "nplurals=2; plural=(n != 1);");